fix: split out screens, implement hardware views, fix paste
