+++ app/screens/onboarding/restore-seed/index.js
@@ -20,6 +20,7 @@ import {
   doRefreshData
 } from "@stores/actions/wallet";
 import { mnemonicToStxAddress, emptySeedArray } from "@utils/utils";
+import { getSeedFromAnyString } from "@common/utils";
 
 const Title = ({ ...rest }) => (
   <Type
@@ -174,7 +175,7 @@ class RestoreSeedScreen extends Component {
   handleOnPaste = (event, index) => {
     if (index === 0) {
       const pasted = event.clipboardData.getData("Text");
-      const split = pasted.trim().split(" ");
+      const split = getSeedFromAnyString(pasted);
       if (split.length === 24) {
         split.forEach((word, i) => {
           this.handleInputChange(
